ISSUE #3584 - fix ticket chip stories after component changes
