+++ frontend/stories/ticketChips/Chip.stories.tsx
@@ -1,0 +1,88 @@
+/**
+ *  Copyright (C) 2022 3D Repo Ltd
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Affero General Public License as
+ *  published by the Free Software Foundation, either version 3 of the
+ *  License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import { ComponentStory, ComponentMeta } from '@storybook/react';
+import { Chip } from '@controls/chip';
+import StarIcon from '@assets/icons/outlined/star-outlined.svg';
+import ClockIcon from '@assets/icons/clock.svg';
+import BellIcon from '@assets/icons/notifications.svg';
+import { ChipStyleWrapper } from './chips.styles';
+
+const Icons = {
+	star: <StarIcon />,
+	clock: <ClockIcon />,
+	bell: <BellIcon />,
+	none: <></>,
+};
+
+// More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
+export default {
+	title: 'Ticket Chips/Chip',
+	component: Chip,
+	// More on argTypes: https://storybook.js.org/docs/react/api/argtypes
+	argTypes: {
+		variant: {
+			description: 'Variant of the chip',
+			options: ['filled', 'outlined', 'text'],
+			control: { type: 'select' },
+		},
+		color: {
+			description: 'The colour of the chip',
+			control: { type: 'color' },
+		},
+		label: {
+			description: 'The text that appears inside the chip',
+			control: { type: 'text' },
+			defaultValue: 'Treatment',
+		},
+		icon: {
+			options: Object.keys(Icons), // An array of serializable values
+			mapping: Icons, // Maps serializable option values to complex arg values
+			control: {
+				type: 'select',
+			},
+		},
+	},
+	parameters: { controls: { exclude: ['size', '$coloroverride', '$variant', 'onDelete', 'deleteIcon', 'avatar', 'sx', 'classes', 'clickable', 'children'] } },
+} as ComponentMeta<typeof Chip>;
+
+// More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
+const Template: ComponentStory<typeof Chip> = (args) => (
+	<ChipStyleWrapper>
+		<div id="tickets">
+			<Chip {...args} />
+		</div>
+	</ChipStyleWrapper>
+);
+
+export const Filled = Template.bind({});
+Filled.args = {
+	variant: 'filled',
+	color: '#00C1D4',
+};
+
+export const Text = Template.bind({});
+Text.args = {
+	variant: 'text',
+	color: '#172B4D',
+};
+
+export const Outlined = Template.bind({});
+Outlined.args = {
+	variant: 'outlined',
+	color: 'hotpink',
+};
